requirements.txt
flask-swagger==0.2.13
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
graphviz==0.14.1
gunicorn==19.9.0
hyperlink==19.0.0
Flask-AppBuilder==2.3.4
certifi==2020.11.8
cffi==1.14.4
idna==2.10
incremental==17.5.0
ipykernel==5.3.3
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
iso8601==0.1.12
isort==4.3.21
itemadapter==0.1.0
itsdangerous==1.1.0
jedi==0.17.2
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
Flask==1.1.2
Jinja2==2.11.2
vine==1.3.0
w3lib==1.22.0
wasabi==0.7.1
wcwidth==0.2.5
webencodings==0.5.1
MarkupSafe==1.1.1
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
timedelta==0.0.0
urllib3==1.26.2
Werkzeug==1.0.1
python-decouple==3.1
